tinyrl: allow same lines, but not in succession, skip empty on save
Principle of least surprise, users coming from other UNIX shells expect
history to be allowed to include repeating lines, as long as it's not in
succession, i.e., this is allowed:

     foo
     bar
     foo
     bar

but not:

     foo
     bar
     foo
     foo

The latter should be compressed to:

     foo
     bar
     foo

Also, never save empty lines to the history.  Very confusing to get back
to the CLI and hitting an empty slot when pressing Up.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>