Refactor script plugin to drop privileges to logged in user
This refactor swaps the roles of the parent and child to be able to drop
privileges from root to the logged in user before running the script.

For root level access, use 'sudo' in the XML file(s), e.g. by setting up
group permissions for some users with NOPASSWD.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>